+++ src/views/KanbanView.ts
@@ -4,7 +4,7 @@ import type { Task, TaskStatus, FilterState, ResolvedProjectConfig } from '../ty
 import type { ProjectScope } from '../store'
 import { flattenTasks, totalLoggedHours } from '../store/TaskTreeOps'
 import { matchesFilter } from '../store/TaskFilter'
-import { dueUrgency, getPriorityConfig } from '../utils'
+import { dueUrgency, getPriorityConfig, safeAsync } from '../utils'
 import { openTaskModal } from '../ui/ModalFactory'
 import { buildTaskContextMenu } from '../ui/TaskContextMenu'
 import { KanbanColumn, type KanbanCardData } from '../ui/composites/KanbanColumn'
@@ -109,6 +109,7 @@ export class KanbanView implements SubView {
       parentTitle,
       projectTitle: owner?.title,
       projectColor: owner?.color,
+      onProjectClick: owner ? safeAsync(() => this.plugin.router.openProjectOverview(owner.filePath)) : undefined,
       loggedHours: totalLoggedHours(task),
       overdue: dueUrgency(task, this.config.statuses) === 'overdue',
       showTagColors: this.plugin.settings.showTagColors